+++ stencil-workspace/src/components.d.ts
@@ -425,9 +425,17 @@ export namespace Components {
         "headerText": string;
         "open": () => Promise<void>;
         /**
+          * (optional) The modal's primary button aria-label.
+         */
+        "primaryButtonAriaLabel": string | null;
+        /**
           * (optional) The modal's primary button text.
          */
         "primaryButtonText": string;
+        /**
+          * (optional) The modal's secondary button aria-label.
+         */
+        "secondaryButtonAriaLabel": string | null;
         /**
           * (optional) The modal's secondary button text.
          */
@@ -1913,9 +1921,17 @@ declare namespace LocalJSX {
          */
         "onSecondaryButtonClick"?: (event: ModusModalCustomEvent<any>) => void;
         /**
+          * (optional) The modal's primary button aria-label.
+         */
+        "primaryButtonAriaLabel"?: string | null;
+        /**
           * (optional) The modal's primary button text.
          */
         "primaryButtonText"?: string;
+        /**
+          * (optional) The modal's secondary button aria-label.
+         */
+        "secondaryButtonAriaLabel"?: string | null;
         /**
           * (optional) The modal's secondary button text.
          */